+++ core/dbt/cli/requires.py
@@ -1,4 +1,5 @@
 import dbt.tracking
+from dbt.common.invocation import reset_invocation_id
 from dbt.version import installed as installed_version
 from dbt.adapters.factory import adapter_management, register_adapter
 from dbt.flags import set_flags, get_flag_dict
@@ -14,9 +15,8 @@ from dbt.common.events.base_types import EventLevel
 from dbt.common.events.functions import (
     fire_event,
     LOG_VERSION,
-    set_invocation_id,
-    setup_event_logger,
 )
+from dbt.events.logging import setup_event_logger
 from dbt.common.events.types import (
     CommandCompleted,
     MainReportVersion,
@@ -52,9 +52,11 @@ def preflight(func):
         ctx.obj["flags"] = flags
         set_flags(flags)
 
+        # Reset invocation_id for each 'invocation' of a dbt command (can happen multiple times in a single process)
+        reset_invocation_id()
+
         # Logging
         callbacks = ctx.obj.get("callbacks", [])
-        set_invocation_id()
         setup_event_logger(flags=flags, callbacks=callbacks)
 
         # Tracking
