+++ core/dbt/common/utils/formatting.py
@@ -1,0 +1,8 @@
+from typing import Optional
+
+
+def lowercase(value: Optional[str]) -> Optional[str]:
+    if value is None:
+        return None
+    else:
+        return value.lower()
