+++ core/dbt/contracts/util.py
@@ -10,7 +10,8 @@ from dbt.exceptions import (
 )
 from dbt.version import __version__
 
-from dbt.common.events.functions import get_invocation_id, get_metadata_vars
+from dbt.common.events.functions import get_metadata_vars
+from dbt.common.invocation import get_invocation_id
 from dbt.common.dataclass_schema import dbtClassMixin
 
 from dbt.common.dataclass_schema import (
