+++ tests/unit/test_functions.py
@@ -2,7 +2,8 @@ from argparse import Namespace
 import pytest
 
 import dbt.flags as flags
-from dbt.common.events.functions import msg_to_dict, warn_or_error, setup_event_logger
+from dbt.common.events.functions import msg_to_dict, warn_or_error
+from dbt.events.logging import setup_event_logger
 from dbt.common.events.types import InfoLevel, NoNodesForSelectionCriteria
 from dbt.common.exceptions import EventCompilationError
